server/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
pydantic>=2.0.0
motor>=3.3.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.0
pydantic[email]>=2.0.0
